+++ src/hci_cmd.c
@@ -585,6 +585,16 @@ const hci_cmd_t hci_write_link_policy_settings = {
 };
 
 /**
+ * @param handle
+ * @param max_latency
+ * @param min_remote_timeout
+ * @param min_local_timeout
+ */
+const hci_cmd_t hci_sniff_subrating = {
+        HCI_OPCODE_HCI_SNIFF_MODE, "H222"
+};
+
+/**
  * @param policy
  */
 const hci_cmd_t hci_write_default_link_policy_setting = {